+++ packages/react-charting/src/components/AreaChart/AreaChart.base.tsx
@@ -27,6 +27,12 @@ import {
   tooltipOfXAxislabels,
   getNextColor,
   getColorFromToken,
+  findNumericMinMaxOfY,
+  createNumericYAxis,
+  IDomainNRange,
+  domainRangeOfNumericForAreaChart,
+  domainRangeOfDateForAreaLineVerticalBarChart,
+  createStringYAxis,
 } from '../../utilities/index';
 import { ILegend, Legends } from '../Legends/index';
 import { DirectionalHint } from '@fluentui/react/lib/Callout';
@@ -150,7 +156,7 @@ export class AreaChartBase extends React.Component<IAreaChartProps, IAreaChartSt
 
   public render(): JSX.Element {
     if (!this._isChartEmpty()) {
-      const { lineChartData, chartTitle } = this.props.data;
+      const { lineChartData } = this.props.data;
       const points = this._addDefaultColors(lineChartData);
       const { colors, opacity, stackedInfo, calloutPoints } = this._createSet(points);
       this._calloutPoints = calloutPoints;
@@ -182,17 +188,21 @@ export class AreaChartBase extends React.Component<IAreaChartProps, IAreaChartSt
       return (
         <CartesianChart
           {...this.props}
-          chartTitle={chartTitle}
+          chartTitle={this._getChartTitle()}
           points={points}
           chartType={ChartTypes.AreaChart}
           calloutProps={calloutProps}
           legendBars={legends}
+          createYAxis={createNumericYAxis}
           isCalloutForStack
           xAxisType={isXAxisDateType ? XAxisTypes.DateAxis : XAxisTypes.NumericAxis}
           tickParams={tickParams}
           maxOfYVal={stackedInfo.maxOfYVal}
           getGraphData={this._getGraphData}
+          getDomainNRangeValues={this._getDomainNRangeValues}
+          createStringYAxis={createStringYAxis}
           getmargins={this._getMargins}
+          getMinMaxOfYAxis={findNumericMinMaxOfY}
           customizedCallout={this._getCustomizedCallout()}
           onChartMouseLeave={this._handleChartMouseLeave}
           enableFirstRenderOptimization={this.props.enablePerfOptimization && this._firstRenderOptimization}
@@ -232,6 +242,35 @@ export class AreaChartBase extends React.Component<IAreaChartProps, IAreaChartSt
       />
     );
   }
+
+  private _getDomainNRangeValues = (
+    points: ILineChartPoints[],
+    margins: IMargins,
+    width: number,
+    chartType: ChartTypes,
+    isRTL: boolean,
+    xAxisType: XAxisTypes,
+    barWidth: number,
+    tickValues: Date[] | number[] | undefined,
+  ) => {
+    let domainNRangeValue: IDomainNRange;
+    if (xAxisType === XAxisTypes.NumericAxis) {
+      domainNRangeValue = domainRangeOfNumericForAreaChart(points, margins, width, isRTL);
+    } else if (xAxisType === XAxisTypes.DateAxis) {
+      domainNRangeValue = domainRangeOfDateForAreaLineVerticalBarChart(
+        points,
+        margins,
+        width,
+        isRTL,
+        tickValues! as Date[],
+        chartType,
+        barWidth,
+      );
+    } else {
+      domainNRangeValue = { dStartValue: 0, dEndValue: 0, rStartValue: 0, rEndValue: 0 };
+    }
+    return domainNRangeValue;
+  };
 
   private _getMargins = (margins: IMargins) => {
     this.margins = margins;
@@ -715,7 +754,15 @@ export class AreaChartBase extends React.Component<IAreaChartProps, IAreaChartSt
       if (!this.props.optimizeLargeData || singleStackedData.length === 1) {
         // Render circles for all data points
         graph.push(
-          <g key={`${index}-dots-${this._uniqueIdForGraph}`} d={area(singleStackedData)!} clipPath="url(#clip)">
+          <g
+            key={`${index}-dots-${this._uniqueIdForGraph}`}
+            d={area(singleStackedData)!}
+            clipPath="url(#clip)"
+            role="region"
+            aria-label={`${points[index].legend}, series ${index + 1} of ${points.length} with ${
+              points[index].data.length
+            } data points.`}
+          >
             {singleStackedData.map((singlePoint: IDPointType, pointIndex: number) => {
               const circleId = `${this._circleId}_${index * this._stackedData[0].length + pointIndex}`;
               const xDataPoint = singlePoint.xVal instanceof Date ? singlePoint.xVal.getTime() : singlePoint.xVal;
@@ -921,4 +968,9 @@ export class AreaChartBase extends React.Component<IAreaChartProps, IAreaChartSt
       // which means chart is not empty
     );
   }
+
+  private _getChartTitle = (): string => {
+    const { chartTitle, lineChartData } = this.props.data;
+    return (chartTitle ? `${chartTitle}. ` : '') + `Area chart with ${lineChartData?.length || 0} data series. `;
+  };
 }