+++ packages/react-components/react-aria/src/AriaLiveAnnouncer/useDomAnnounce.ts
@@ -144,8 +144,10 @@ export const useDomAnnounce_unstable = (): AriaLiveAnnounceFn => {
     return () => {
       element.remove();
       elementRef.current = null;
+      clearAnnounceTimeout();
+      timeoutRef.current = undefined;
     };
-  }, [targetDocument]);
+  }, [clearAnnounceTimeout, targetDocument]);
 
   return announce;
 };