+++ packages/react-components/react-carousel-preview/src/components/Carousel/useCarousel.ts
@@ -7,6 +7,8 @@ import {
   useEventCallback,
   useMergedRefs,
 } from '@fluentui/react-utilities';
+import { useFluent_unstable as useFluent } from '@fluentui/react-shared-contexts';
+
 import type { CarouselProps, CarouselState } from './Carousel.types';
 import { useCarouselWalker_unstable } from '../useCarouselWalker';
 import { createCarouselStore } from '../createCarouselStore';
@@ -23,8 +25,10 @@ import type { CarouselContextValue } from '../CarouselContext.types';
  * @param ref - reference to root HTMLDivElement of Carousel
  */
 export function useCarousel_unstable(props: CarouselProps, ref: React.Ref<HTMLDivElement>): CarouselState {
-  const { onValueChange } = props;
+  const { onValueChange, circular } = props;
 
+  const { targetDocument } = useFluent();
+  const win = targetDocument?.defaultView;
   const { ref: carouselRef, walker: carouselWalker } = useCarouselWalker_unstable();
   const [store] = React.useState(() => createCarouselStore());
 
@@ -60,6 +64,10 @@ export function useCarousel_unstable(props: CarouselProps, ref: React.Ref<HTMLDi
   }, [store]);
 
   React.useEffect(() => {
+    if (!win) {
+      return;
+    }
+
     const config: MutationObserverInit = {
       attributes: true,
       attributeFilter: [CAROUSEL_ITEM],
@@ -94,7 +102,7 @@ export function useCarousel_unstable(props: CarouselProps, ref: React.Ref<HTMLDi
     };
 
     // Create an observer instance linked to the callback function
-    const observer = new MutationObserver(callback);
+    const observer = new win.MutationObserver(callback);
 
     // Start observing the target node for configured mutations
     observer.observe(rootRef.current!, config);
@@ -103,7 +111,7 @@ export function useCarousel_unstable(props: CarouselProps, ref: React.Ref<HTMLDi
     return () => {
       observer.disconnect();
     };
-  }, [carouselWalker, store]);
+  }, [carouselWalker, store, win]);
 
   const selectPageByDirection: CarouselContextValue['selectPageByDirection'] = useEventCallback((event, direction) => {
     const active = carouselWalker.active();
@@ -113,7 +121,9 @@ export function useCarousel_unstable(props: CarouselProps, ref: React.Ref<HTMLDi
     }
 
     const newPage =
-      direction === 'prev' ? carouselWalker.prevPage(active.value) : carouselWalker.nextPage(active.value);
+      direction === 'prev'
+        ? carouselWalker.prevPage(active.value, circular)
+        : carouselWalker.nextPage(active.value, circular);
 
     if (newPage) {
       setValue(newPage?.value);
@@ -133,6 +143,7 @@ export function useCarousel_unstable(props: CarouselProps, ref: React.Ref<HTMLDi
     root: slot.always(
       getIntrinsicElementProps('div', {
         ref: useMergedRefs(ref, carouselRef, rootRef),
+        role: 'region',
         ...props,
       }),
       { elementType: 'div' },
@@ -141,5 +152,6 @@ export function useCarousel_unstable(props: CarouselProps, ref: React.Ref<HTMLDi
     value,
     selectPageByDirection,
     selectPageByValue,
+    circular,
   };
 }