+++ packages/react-components/react-carousel-preview/src/components/CarouselButton/CarouselButton.types.ts
@@ -3,7 +3,7 @@ import { ARIAButtonSlotProps } from '@fluentui/react-aria';
 import { ButtonProps, ButtonSlots, ButtonState } from '@fluentui/react-button';
 
 export type CarouselButtonSlots = ButtonSlots & {
-  root: NonNullable<Slot<ARIAButtonSlotProps<'a'>>>;
+  root: NonNullable<Slot<ARIAButtonSlotProps>>;
 };
 
 /**
@@ -11,6 +11,9 @@ export type CarouselButtonSlots = ButtonSlots & {
  */
 export type CarouselButtonProps = ButtonProps &
   ComponentProps<CarouselButtonSlots> & {
+    /**
+     * Dictates whether button will be of type go next or go previous
+     */
     navType: 'prev' | 'next';
   };