+++ packages/react-components/react-carousel-preview/src/components/CarouselNavImageButton/renderCarouselNavImageButton.tsx
@@ -10,6 +10,9 @@ import type { CarouselNavImageButtonState, CarouselNavImageButtonSlots } from '.
 export const renderCarouselNavImageButton_unstable = (state: CarouselNavImageButtonState) => {
   assertSlots<CarouselNavImageButtonSlots>(state);
 
-  // TODO Add additional slots in the appropriate place
-  return <state.root />;
+  return (
+    <state.root>
+      <state.image />
+    </state.root>
+  );
 };