+++ packages/react-components/react-message-bar/stories/MessageBar/Shape.stories.tsx
@@ -1,11 +1,11 @@
 import * as React from 'react';
-import { makeStyles, shorthands, MessageBar, MessageBarTitle, MessageBarBody } from '@fluentui/react-components';
+import { makeStyles, MessageBar, MessageBarTitle, MessageBarBody } from '@fluentui/react-components';
 
 const useClasses = makeStyles({
   container: {
     display: 'flex',
     flexDirection: 'column',
-    ...shorthands.gap('10px'),
+    gap: '10px',
   },
 });