+++ packages/react-components/react-motions-preview/src/factories/createMotionComponent.test.tsx
@@ -6,7 +6,6 @@ import { createMotionComponent } from './createMotionComponent';
 
 const motion: AtomMotion = {
   keyframes: [{ opacity: 0 }, { opacity: 1 }],
-
   duration: 500,
 };
 
@@ -14,6 +13,7 @@ function createElementMock() {
   const animateMock = jest.fn().mockImplementation(() => ({
     play: jest.fn(),
     cancel: jest.fn(),
+    persist: jest.fn(),
   }));
   const ElementMock = React.forwardRef((props, ref) => {
     React.useImperativeHandle(ref, () => ({
@@ -43,7 +43,6 @@ describe('createMotionComponent', () => {
     expect(animateMock).toHaveBeenCalledWith(motion.keyframes, {
       duration: motion.duration,
       fill: 'forwards',
-      iterations: 1,
     });
   });