+++ packages/react-components/react-provider/src/components/FluentProvider/FluentProvider-node.test.tsx
@@ -33,7 +33,7 @@ describe('FluentProvider (node)', () => {
     expect(parseHTMLString(html)).toMatchInlineSnapshot(`
       "<div
         dir="ltr"
-        class="fui-FluentProvider fui-FluentProvider1 "
+        class="fui-FluentProvider fui-FluentProvider1"
       >
         <style id="fui-FluentProvider1">
           .fui-FluentProvider1 {
@@ -60,7 +60,7 @@ describe('FluentProvider (node)', () => {
     expect(parseHTMLString(html)).toMatchInlineSnapshot(`
       "<div
         dir="ltr"
-        class="fui-FluentProvider fui-FluentProvider1 "
+        class="fui-FluentProvider fui-FluentProvider1"
       >
         <style nonce="random" id="fui-FluentProvider1">
           .fui-FluentProvider1 {