+++ packages/react-components/react-provider/stories/Provider/FluentProviderFrame.stories.tsx
@@ -3,7 +3,6 @@ import * as ReactDOM from 'react-dom';
 import {
   createDOMRenderer,
   makeStyles,
-  shorthands,
   tokens,
   Button,
   FluentProvider,
@@ -18,17 +17,17 @@ const useExampleStyles = makeStyles({
     backgroundColor: tokens.colorBrandBackground2,
     color: tokens.colorBrandForeground2,
     fontSize: '20px',
-    ...shorthands.border('1px'),
-    ...shorthands.borderRadius('5px'),
-    ...shorthands.padding('5px'),
+    border: '1px',
+    borderRadius: '5px',
+    padding: '5px',
   },
 });
 
 const useProviderStyles = makeStyles({
   provider: {
-    ...shorthands.border('1px'),
-    ...shorthands.borderRadius('5px'),
-    ...shorthands.padding('5px'),
+    border: '1px',
+    borderRadius: '5px',
+    padding: '5px',
   },
 });